requirements-prod.txt
Django==3.2
django-uuslug==1.2.0
django-widget-tweaks==1.4.8
graphene-django==2.15.0
python-dateutil==2.8.1
user-agents==2.2.0
Pillow==8.2.0
celery==4.4.7
django-celery-email==3.0.0
psycopg2-binary==2.8.6
gunicorn==20.1.0
django-redis==4.12.1
